+++ lib/ApplePayDomain.php
@@ -20,7 +20,8 @@ class ApplePayDomain extends ApiResource
     }
 
     /**
-     * @param string $id The ID of the domain to retrieve.
+     * @param array|string $id The ID of the domain to retrieve, or an options
+     *     array containing an `id` key.
      * @param array|string|null $opts
      *
      * @return ApplePayDomain