+++ lib/CountrySpec.php
@@ -21,7 +21,9 @@ class CountrySpec extends ApiResource
     }
 
     /**
-     * @param string $country The ISO country code of the country we retrieve the CountrySpec for.
+     * @param array|string $country The ISO country code of the country we
+     *     retrieve the country specfication for, or an options array
+     *     containing an `id` containing that code.
      * @param array|string|null $opts
      *
      * @return CountrySpec