+++ lib/Invoice.php
@@ -21,7 +21,8 @@ class Invoice extends ApiResource
     }
 
     /**
-     * @param string $id The ID of the invoice to retrieve.
+     * @param array|string $id The ID of the invoice to retrieve, or an options
+     *     array containing an `id` key.
      * @param array|string|null $opts
      *
      * @return Invoice