+++ lib/InvoiceItem.php
@@ -10,7 +10,8 @@ namespace Stripe;
 class InvoiceItem extends ApiResource
 {
     /**
-     * @param string $id The ID of the invoice item to retrieve.
+     * @param array|string $id The ID of the invoice item to retrieve, or an
+     *     options array containing an `id` key.
      * @param array|string|null $opts
      *
      * @return InvoiceItem