+++ lib/Product.php
@@ -10,7 +10,8 @@ namespace Stripe;
 class Product extends ApiResource
 {
     /**
-     * @param string $id The ID of the Product to retrieve.
+     * @param array|string $id The ID of the product to retrieve, or an options
+     *     array contianing an `id` key.
      * @param array|string|null $opts
      *
      * @return Product