+++ lib/Recipient.php
@@ -10,7 +10,8 @@ namespace Stripe;
 class Recipient extends ApiResource
 {
     /**
-     * @param string $id The ID of the recipient to retrieve.
+     * @param array|string $id The ID of the recipient to retrieve, or an
+     *     options array containing an `id` key.
      * @param array|string|null $opts
      *
      * @return Recipient