+++ lib/Stripe.php
@@ -30,7 +30,11 @@ class Stripe
     // @var array The application's information (name, version, URL)
     public static $appInfo = null;
 
-    const VERSION = '4.5.1';
+    // @var Util\LoggerInterface|null The logger to which the library will
+    //   produce messages.
+    public static $logger = null;
+
+    const VERSION = '4.9.1';
 
     /**
      * @return string The API key used for requests.
@@ -38,6 +42,27 @@ class Stripe
     public static function getApiKey()
     {
         return self::$apiKey;
+    }
+
+    /**
+     * @return Util\LoggerInterface The logger to which the library will
+     *   produce messages.
+     */
+    public static function getLogger()
+    {
+        if (self::$logger == null) {
+            return new Util\DefaultLogger();
+        }
+        return self::$logger;
+    }
+
+    /**
+     * @param Util\LoggerInterface $logger The logger to which the library
+     *   will produce messages.
+     */
+    public static function setLogger($logger)
+    {
+        self::$logger = $logger;
     }
 
     /**