+++ lib/Token.php
@@ -20,7 +20,8 @@ namespace Stripe;
 class Token extends ApiResource
 {
     /**
-     * @param string $id The ID of the token to retrieve.
+     * @param array|string $id The ID of the token to retrieve, or an options
+     *     array containing an `id` key.
      * @param array|string|null $opts
      *
      * @return Token